requirements.txt
anthropic==0.86.0
alembic==1.18.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.31.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
cachetools==5.5.2
camelot-py==1.0.9
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dash==3.2.0
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.2
et_xmlfile==2.0.0
fastapi==0.95.0
Flask==2.0.1
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==1.48.1
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.97.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
plotly==6.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.4.2
rpds-py==0.26.0
rsa==4.9.1
seaborn==0.13.2
selenium==4.32.0
setuptools==80.9.0
six==1.17.0
slack_sdk==3.36.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
standardwebhooks==1.0.1
starlette==0.26.1
streamlit==1.47.1
svix==1.89.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-Deprecated==1.3.1.20260130
types-python-dateutil==2.9.0.20260323
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.21.1
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==2.0.1
wheel==0.45.1
wrapt==2.1.2
wsproto==1.2.0
zipp==3.23.0
